add entity

+++ src/main/java/repository/ProductRepo.java
@@ -2,12 +2,13 @@ package repository;
 
 import entity.Product;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ProductRepo {
     List<Product> findAll();
     Product findById(Integer id);
-    void createProduct (String name, Double price);
-    void updateProduct(Integer id, String name, Double price);
+    void createProduct (String name, BigDecimal price);
+    void updateProduct(Integer id, String name, BigDecimal price);
     void deleteProductById(Integer id);
 }
